Creates shared carbon-models array
