+++ co2eq/index.test.js
@@ -1,24 +1,5 @@
-import * as electricityWorldAverageCarbonModel from './electricity/worldaverage';
-import * as mealCarbonModel from './food/meal';
-import * as ingredientsCarbonModel from './food/ingredients';
-import * as transportationCarbonModel from './transportation/index';
-import * as flightCarbonModel from './flights/index';
-import * as carCarbonModel from './car/index';
-import * as purchaseCarbonModel from './purchase/index';
-import * as energyCarbonModel from './energy/index';
-import * as hotelCarbonModel from './hotelpercountry/index';
+import carbonModels from './index';
 
-const carbonModels = [
-  purchaseCarbonModel,
-  electricityWorldAverageCarbonModel,
-  ingredientsCarbonModel,
-  mealCarbonModel,
-  carCarbonModel,
-  transportationCarbonModel,
-  flightCarbonModel,
-  energyCarbonModel,
-  hotelCarbonModel,
-];
 describe(`all models have valid API`, () => {
   carbonModels.forEach(model => {
     test(`${model.modelName}`, () => {
